+++ src/features/studyInfo/LeftSection.tsx
@@ -41,8 +41,8 @@ const hrStyle = css`
 `;
 
 const studyDescriptionStyle = css`
-    font-size: 12px;
-    line-height: 15px;
+    font-size: 15px;
+    line-height: 18px;
     color: white;
     font-weight: lighter;
     margin-top: 20px;
