+++ src/features/studyInfo/RightSection.tsx
